Adjusted Numbering (label) of LaBr detectors
Adjusted the Numbering of some histograms of the LaBr detectors, such that the detectors number (down in hte lab) should correspond to the histogram name, eg. energy_labr_01. So 01 is the first detector.

For the E detectors and dE I left it with the previous convention, sthat they start from 0 (at least for now).